QA ADD REMARKS Function

+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/qa/QaRemarksEntity.kt
@@ -33,7 +33,7 @@ class QaRemarksEntity : Serializable {
 
     @Column(name = "REMARKS_STATUS")
     @Basic
-    var remarksStatus: Long? = null
+    var remarksStatus: Int? = null
 
     @Column(name = "DESCRIPTION")
     @Basic
